Move DS to module (#337)
