+++ data/src/lib.rs
@@ -18,6 +18,7 @@ pub mod servers;
 pub mod transaction_store;
 pub mod mp_store;
 pub mod observation_store;
+pub mod data_store;
 
 #[derive(Clone)]
 pub struct DataStoreContext {
