+++ src/core/peer_rx_event_handler.rs
@@ -23,7 +23,7 @@ use crate::core::discovery::DiscoveryMessage;
 // use crate::api::p2p_io::rgnetwork::{Client, Event, PeerResponse};
 use crate::core::internal_message::{new_channel, PeerMessage, RecvAsyncErrorInfo, SendErrorInfo, TransactionMessage};
 use crate::core::relay::{ Relay};
-use crate::data::data_store::DataStore;
+use redgold_data::data_store::DataStore;
 use crate::data::download::process_download_request;
 use crate::multiparty::initiate_mp::{initiate_mp_keygen, initiate_mp_keygen_follower, initiate_mp_keysign, initiate_mp_keysign_follower};
 use crate::node_config::NodeConfig;
@@ -151,7 +151,7 @@ impl PeerRxEventHandler {
 
         if let Some(download_request) = request.download_request {
             // info!("Received download request");
-            let result = DataStore::map_err(process_download_request(&relay, download_request).await)?;
+            let result = process_download_request(&relay, download_request).await?;
             response.download_response = Some(result);
         }
