+++ src/data/mod.rs
@@ -1,4 +1,2 @@
-pub mod block_store;
-pub mod data_store;
 pub mod download;
 pub mod utxo;
